move EditableMeta to QueryResult
