+++ src/Core/Domain/Meta/QueryHandler/GetMetaForEditingHandlerInterface.php
@@ -26,7 +26,7 @@
 
 namespace PrestaShop\PrestaShop\Core\Domain\Meta\QueryHandler;
 
-use PrestaShop\PrestaShop\Core\Domain\Meta\EditableMeta;
+use PrestaShop\PrestaShop\Core\Domain\Meta\QueryResult\EditableMeta;
 use PrestaShop\PrestaShop\Core\Domain\Meta\Query\GetMetaForEditing;
 
 /**
